requirements.txt
# Juriitech — Python-afhængigheder
# Streamlit Cloud læser denne fil og installerer pakkerne automatisk.

# Streamlit UI
streamlit>=1.42.0

# Database (Neon Postgres + pgvector)
psycopg2-binary>=2.9.9
pgvector>=0.4.0

# Miljøvariabler
python-dotenv>=1.0.0

# Anthropic Claude (AI)
anthropic>=0.45.0

# Voyage AI (embeddings)
voyageai>=0.3.0

# Filhåndtering (PDF / Word)
pypdf>=4.0.0
python-docx>=1.1.0

# Web scraping (Ankenævn + TUI)
requests>=2.31.0
beautifulsoup4>=4.12.0

# Authlib — OIDC/OAuth til login (Microsoft, Google)
Authlib>=1.3.2
